Add metawatcher (#13)
Add meta watcher to prevent restart stalls

When stale data is detected, we close the current connection and attempt to establish a new one. If that failed, the app used to gets wedged. Now, a watcher is started that checks if the connection was established successfully, and retries every minute until it does.